requirements.txt

fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
langchain>=0.0.335
langchain-openai>=0.0.2
openai>=1.2.0
python-multipart>=0.0.6
pypdf>=3.17.0
requests>=2.31.0
qdrant-client>=1.6.4
psycopg2-binary>=2.9.9
pymongo>=4.5.0
pydantic-settings>=2.8.1